feat(stella-core): warn the agent when a turn ends with a declared service still up
#2666 fixed the two mechanical causes of "correct service, agent proves it
works, verifier finds it dead": start_process children SIGKILLed on
ProcessTable::drop, and bash hung on a backgrounded grandchild. It left the
third, which is not mechanical — nothing ever asked the agent about the
service before the turn stopped.

A run can now leave a service up correctly, so "still running" is no longer
evidence of a mistake, and that is exactly why it needs saying rather than
acting on. Two very different turns end identically: the service IS the
deliverable, or the model started something, never checked it was reachable,
and declared done anyway. Only the model knows which. So `driver::live_services`
states the fact and hands the decision back — one nudge per turn, naming every
still-running handle, on the prove-it gate's mechanism (#2663) and with its
bound. It stops nothing: stop_process remains the only path that kills a
started child, because a surviving declared service can be the correct final
state.

The engine holds no process table and must not (invariant 1), so the answer
comes through a new non-destructive `ToolExecutor::live_services` — a peek,
not a drain, because it reports state outliving the turn rather than a ledger
the first caller settles. `start_process` answers it from the table; the
registry aggregates across tools, oldest handle first so what reaches the
prompt is deterministic (invariant 7); every decorator forwards, pinned
end-to-end by `the_production_tool_stack_forwards_live_services`. The three
executors that keep the empty default (serve's remote port, the witness
author's surface, and the read-only view's deliberate non-filter) each say why.

Both once-per-turn gates now share one nudge window: two gates each appending
a user message would otherwise reopen each other's, reproducing the `gate-ab`
regression (#2663) across gates instead of within one.

To keep three god files closed to growth, `registry.rs`'s ToolExecutor impl
moved to `registry/executor.rs` and `custom.rs`'s test module to
`custom/tests.rs` (the media.rs/process.rs precedent). The one baseline cell
that rose is `driver/tests.rs` +1, for a `mod` declaration; `registry.rs`'s
ceiling is retightened to its new size so the split buys structure, not slack.

Closes #2764
Refs #2666, #2663, #2811, #2818, #2819
