+++ crates/stella-core/src/driver/tests.rs
@@ -3085,6 +3085,7 @@ mod compute_passes;
 mod context_efficiency;
 mod context_overflow;
 mod lifecycle_bus;
+mod live_services;
 mod loop_abort;
 mod model_fallback;
 mod parked_wait;
