+++ crates/stella-pipeline/README.md
@@ -102,6 +102,7 @@ as a planning assumption.
 | [`src/candidate.rs`](src/candidate.rs) | `CandidateScore` and `select_best_candidate` — best-of-N selection, pure. |
 | [`src/candidate_fanout.rs`](src/candidate_fanout.rs) | `fan_out_width` and `FanOutBudget` — how wide a fan-out runs and how one turn's money is split between candidates spending at the same time. Pure; the normative statement of the overshoot window. |
 | [`src/pipeline/fanout_stage.rs`](src/pipeline/fanout_stage.rs) | The concurrent dispatch itself: workspace creation (serialized), `buffer_unordered` over the candidates, results re-ordered by index. |
+| [`src/pipeline/attachments.rs`](src/pipeline/attachments.rs), [`src/pipeline/fallback.rs`](src/pipeline/fallback.rs) | The single funnel every engine this pipeline builds goes through (gate, calibration, breaker feedback, mid-turn fallback), and the router-backed `FallbackResolver` it attaches. Every engine site states a `FallbackPosture` — re-resolve this role, or a named reason for withholding. |
 | [`src/mcp_prefetch.rs`](src/mcp_prefetch.rs) | `fold`: shared MCP context gathered once at the top of a fan-out instead of N times. |
 | [`src/replay.rs`](src/replay.rs), [`src/replay/golden.rs`](src/replay/golden.rs) | `validate_stream` / `structural_diff` / `parse_jsonl`, and the golden fixture format with its provenance manifest. |
