+++ crates/stella-tools/README.md
@@ -83,7 +83,7 @@ tool's registration line in an already-oversized `registry.rs`), never as a plan
 | File | What it holds |
 |---|---|
 | [`src/lib.rs`](src/lib.rs) | Module list plus `resolve_within_root` — the one path-confinement primitive every tool resolves through. |
-| [`src/registry.rs`](src/registry.rs), [`src/registry/process_tools.rs`](src/registry/process_tools.rs) | The `Tool` trait, `ToolRegistry`, construction (which tools register under which conditions), and the single `execute` path all cross-cutting behaviour hangs off. `process_tools.rs` is the closed list of child-process-spawning built-ins. |
+| [`src/registry.rs`](src/registry.rs), [`src/registry/process_tools.rs`](src/registry/process_tools.rs), [`src/registry/executor.rs`](src/registry/executor.rs) | The `Tool` trait, `ToolRegistry`, construction (which tools register under which conditions), and the single `execute` path all cross-cutting behaviour hangs off. `process_tools.rs` is the closed list of child-process-spawning built-ins; `executor.rs` is the `ToolExecutor` port impl (`schemas`/`execute` plus the drains and aggregations the engine reads). |
 | [`src/catalog.rs`](src/catalog.rs) | The canonical tool table. Open it to add a tool or to answer "is this name taken / is it read-only". |
 | [`src/verify.rs`](src/verify.rs) | `verify_done` — the shadow-worktree witness gate. See below. |
 | [`src/read.rs`](src/read.rs), [`src/read_symbol.rs`](src/read_symbol.rs), [`src/write.rs`](src/write.rs), [`src/edit.rs`](src/edit.rs), [`src/apply_edits.rs`](src/apply_edits.rs), [`src/delete.rs`](src/delete.rs) | File CRUD. They share one read-state ledger so an `old_string` miss can be attributed to out-of-band drift rather than to the model. |
@@ -94,7 +94,7 @@ tool's registration line in an already-oversized `registry.rs`), never as a plan
 | [`src/exec.rs`](src/exec.rs) | The shared subprocess runner: process-group spawn, hard timeout with group kill, output truncation, and the env-scrub constants. |
 | [`src/subprocess_env.rs`](src/subprocess_env.rs) | The credential deny-list applied as the last env mutation before any model- or repo-controlled spawn. Downstream crates use this, never a copy. |
 | [`src/project.rs`](src/project.rs), [`src/scripts.rs`](src/scripts.rs), [`src/diagnostics.rs`](src/diagnostics.rs), [`src/impact.rs`](src/impact.rs) | Build/test/lint/format as thin verbs over the scripts index, structured typecheck output, and the importer-edge blast radius behind `run_tests` `scope: "impacted"`. |
-| [`src/process.rs`](src/process.rs) | The long-running process group (`start_process`/`read_output`/`send_stdin`/`stop_process`) for servers, REPLs, watchers. |
+| [`src/process.rs`](src/process.rs) | The long-running process group (`start_process`/`read_output`/`send_stdin`/`stop_process`) for servers, REPLs, watchers. Also answers `Tool::live_services`, the read behind the engine's end-of-turn assertion (#2764). |
 | [`src/repo.rs`](src/repo.rs), [`src/ci.rs`](src/ci.rs) | Vendor-neutral `repo_*` tools behind the `RepoBackend` port (`GitCli` is the only adapter), and `ci_status` via `gh`. |
 | [`src/bash.rs`](src/bash.rs) | The default-on shell, withheld with `"tools": {"bash": "off"}`. No per-command OS confinement: the `STELLA_BASH_SANDBOX` Seatbelt/`bwrap` wrapper was removed in #1300 because it bounded this one tool while every other spawn path went around it. |
 | [`src/web.rs`](src/web.rs), [`src/web_extract.rs`](src/web_extract.rs) | The opt-in web family, and the pure HTML/CSS extraction behind it (no I/O in `web_extract.rs`, so it unit-tests without a network). |
